feat: resolve rebase conflicts with main
