+++ content/blog/dojo-kun.md
