+++ pages/index.vue
@@ -9,11 +9,18 @@ const { data: articles } = await useAsyncData('equal', () =>
 <template>
   <div>
     <SectionsHeader />
-    <ul class="container">
-      <li v-for="article of articles" :key="article.id" class="article-card">
-        <SectionsBlogCard :article="article" />
-      </li>
-    </ul>
+    <div class="main">
+      <h1 class="h1">Entradas recientes</h1>
+      <ul class="container">
+        <li
+          v-for="article of articles"
+          :key="article.slug"
+          class="article-card"
+        >
+          <SectionsBlogCard :article="article" />
+        </li>
+      </ul>
+    </div>
 
     <SectionsFooter />
   </div>
